exception fix
git-svn-id: http://svn.l2jfree.com/l2jfree/branches/genesis@9352 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ l2jfree-core/src/main/java/com/l2jfree/gameserver/sql/PlayerDB.java
@@ -250,7 +250,7 @@ public class PlayerDB extends L2DBEntity
 			public void configure(Query q)
 			{
 				q.setParameter("online", isOnline);
-				q.setParameter("persistentId", player.getPersistentId());
+				q.setParameter("persistentId", player.getPersistentId().intValue());
 			}
 		});
 	}
